Refactor: UI 리팩토링/최적화
Inventory UI Content 제거 및 Inventory UI로 통합
아이템 슬롯 정렬 방식을 Bubble Sort -> 투 포인터 전략으로 변경, 최적화 진행
