+++ Content/Blueprints/02_AbilitySystem/01_Player/AS_AkaOni.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:6ad4586bc58ab50c38ec940ee7b8605d6c482a03b88763e8be9381d38f010925
-size 21054
+oid sha256:cab9a30d686004b4efbf0855142afaed4e38b76bf53059ef201a4500d2ac452a
+size 21270
