+++ Content/StarterContent/Audio/Collapse01.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:85c827023ee4b466228ec65e92b4b4a0415a106de041ac1c2cab481a2a5dfa8d
-size 348532
+oid sha256:2310e2f199c641934c290bd565ff76c9b492a01f8ba9ba901fe4fcb769be4730
+size 353652
