+++ Source/AgeOfWolves/08_UI/01_HUD/HUD_StatusUI.h
@@ -1,5 +1,3 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
 #pragma once
 
 #include "CoreMinimal.h"
