+++ Source/AgeOfWolves/08_UI/02_Menu/01_InventoryUI/InventoryUI.cpp
@@ -4,8 +4,10 @@
 #include "Components/Overlay.h"
 #include "Components/OverlaySlot.h"
 
-#include "08_UI/02_Menu/01_InventoryUI/InventoryUIContent.h"
+#include "08_UI/02_Menu/01_InventoryUI/InventoryToolBar.h"
 #include "08_UI/02_Menu/01_InventoryUI/ItemSlots.h"
+#include "08_UI/02_Menu/01_InventoryUI/ItemDescriptionSlot.h"
+#include "08_UI/InteractableItemSlot.h"
 
 DEFINE_LOG_CATEGORY(LogInventoryUI)
 
@@ -16,6 +18,7 @@ UInventoryUI::UInventoryUI(const FObjectInitializer& ObjectInitializer)
 {
     //@Menu Category
     MenuCategory = EMenuCategory::Inventory;
+    InventoryToolBar = nullptr;
 }
 
 void UInventoryUI::NativeOnInitialized()
@@ -36,6 +39,8 @@ void UInventoryUI::NativePreConstruct()
 void UInventoryUI::NativeConstruct()
 {
     Super::NativeConstruct();
+
+    SetIsFocusable(true);
 }
 
 void UInventoryUI::NativeDestruct()
@@ -43,54 +48,266 @@ void UInventoryUI::NativeDestruct()
     Super::NativeDestruct();
 }
 
-FReply UInventoryUI::NativeOnFocusReceived(const FGeometry& InGeometry, const FFocusEvent& InFocusEvent)
+FNavigationReply UInventoryUI::NativeOnNavigation(const FGeometry& MyGeometry, const FNavigationEvent& InNavigationEvent, const FNavigationReply& InDefaultReply)
 {
-    FReply Reply = Super::NativeOnFocusReceived(InGeometry, InFocusEvent);
-
-    //@SetDirectly(SetFocus())에 의한 포커스가 아닐 경우, 포커스를 해제합니다.
-    if (InFocusEvent.GetCause() != EFocusCause::SetDirectly)
+    //@Navigation 방향에 따라 처리
+    switch (InNavigationEvent.GetNavigationType())
     {
-        return Reply.Handled().ClearUserFocus();
-    }
-
-    if (InFocusEvent.GetCause() == EFocusCause::SetDirectly && Reply.IsEventHandled())
+    case EUINavigation::Down:
     {
-        if (!InventoryUIContent)
+        //@아래 방향키: ItemSlots로 포커스 이동
+        UItemSlots* CurrentItemSlots = Cast<UItemSlots>(GetItemSlotsUI(CurrentItemType));
+        if (CurrentItemSlots)
         {
-            UE_LOGFMT(LogInventoryUI, Warning, "InventoryUIContent가 유효하지 않습니다. 포커스를 설정할 수 없습니다.");
-            return Reply;
-        }
+            //@첫 번째 아이템 슬롯 유효한가?
+            UInteractableItemSlot* FirstItemSlot = CurrentItemSlots->FindFirstItemSlot();
+            if (FirstItemSlot && FirstItemSlot->GetUniqueItemID().IsValid())
+            {
+                //@첫 번째 아이템 슬롯의 강제 호버 상태 전환 요청 이벤트
+                RequestFirstItemSlotHover.Broadcast(CurrentItemType);
 
-        //@Set Focus
-        InventoryUIContent->SetFocus();
+                //@SetFocus
+                CurrentItemSlots->SetFocus();
+
+                UE_LOGFMT(LogInventoryUI, Log, "Navigation Down: 포커스가 {0} 타입의 ItemSlots로 이동했습니다.", *UEnum::GetValueAsString(CurrentItemType));
+
+                //@SetFocus를 통해 포커스를 주었으므로 nullptr을 반환
+                return FNavigationReply::Explicit(nullptr);
+            }
+        }
+        break;
+    }
+    case EUINavigation::Up:
+    {
+        //@위 방향키: ItemSlots에서 InventoryToolBar로 포커스 이동
+        UItemSlots* CurrentItemSlots = Cast<UItemSlots>(GetItemSlotsUI(CurrentItemType));
+        if (CurrentItemSlots && CurrentItemSlots->HasKeyboardFocus())
+        {
+            //@Request Cancel Current Hovered Item Slot
+            RequestCancelCurrentHoveredItemSlot.Broadcast(CurrentItemType);
+
+            //@ToolBar에게 포커스 전달
+            if (InventoryToolBar)
+            {
+                InventoryToolBar->SetFocus();
+                UE_LOGFMT(LogInventoryUI, Log, "Navigation Up: InventoryToolBar에게 포커스를 전달합니다.");
+                //@SetFocus를 통해 포커스를 주었으므로 nullptr을 반환
+                return FNavigationReply::Explicit(nullptr);
+            }
+        }
+        break;
+    }
+    case EUINavigation::Left:
+    case EUINavigation::Right:
+    {
+        //@좌우 방향키는 ToolBar에서 처리
+        if (InventoryToolBar && InventoryToolBar->HasKeyboardFocus())
+        {
+            //@Navigation을 ToolBar로 전달
+            return InDefaultReply;
+        }
+        break;
+    }
+    default:
+        break;
     }
 
-    return Reply;
+    //@기본적으로 Navigation 차단
+    return FNavigationReply::Explicit(nullptr);
 }
 
-void UInventoryUI::InternalBindingToInventoryUIContent(UInventoryUIContent* Content)
+FReply UInventoryUI::NativeOnFocusReceived(const FGeometry& InGeometry, const FFocusEvent& InFocusEvent)
 {
-    //@Item Description
-    if (!Content)
+    //@SetDirectly(SetFocus())를 통한 포커스 시도 외에 다른 시도는 허용하지 않습니다.
+    if (InFocusEvent.GetCause() != EFocusCause::SetDirectly)
     {
-        UE_LOGFMT(LogInventoryUI, Error, "InventoryUIContent UI가 유효하지 않습니다.");
+        return FReply::Handled().ClearUserFocus();
+    }
+
+    UE_LOGFMT(LogInventoryUI, Log, "포커스 : 위젯: {0}, 원인: {1}",
+        *GetName(), *UEnum::GetValueAsString(InFocusEvent.GetCause()));
+
+    //@최초 포커스 시 InventoryToolBar에게 포커스 전달
+    if (InventoryToolBar)
+    {
+        //@ToolBar에게 포커스 전달
+        InventoryToolBar->SetFocus();
+        UE_LOGFMT(LogInventoryUI, Log, "InventoryUI가 포커스를 받았습니다. InventoryToolBar에게 포커스를 전달합니다.");
+    }
+    else
+    {
+        UE_LOGFMT(LogInventoryUI, Warning, "InventoryToolBar가 유효하지 않아 포커스를 전달할 수 없습니다.");
+    }
+
+    return FReply::Handled();
+}
+
+void UInventoryUI::NativeOnFocusLost(const FFocusEvent& InFocusEvent)
+{
+    //@SetDirectly(SetFocus())를 통한 포커스 소실 외에 다른 시도는 허용하지 않습니다.
+    if (InFocusEvent.GetCause() != EFocusCause::SetDirectly)
+    {
+        SetFocus();
         return;
     }
 
-    Content->InventoryUIContentInitFinished.BindUFunction(this, "OnInventoryUIContentInitFinished");
+    Super::NativeOnFocusLost(InFocusEvent);
+
+    UE_LOGFMT(LogInventoryUI, Log, "포커스 종료: 위젯: {0}, 원인: {1}",
+        *GetName(), *UEnum::GetValueAsString(InFocusEvent.GetCause()));
+}
+
+FReply UInventoryUI::NativeOnKeyDown(const FGeometry& InGeometry, const FKeyEvent& InKeyEvent)
+{
+    FKey Key = InKeyEvent.GetKey();
+
+    UE_LOGFMT(LogInventoryUI, Log, "키 입력 감지됨: {0}", *Key.ToString());
+
+    //@Inventory Tool Bar
+    if (!InventoryToolBar)
+    {
+        UE_LOGFMT(LogInventoryUI, Error, "InventoryToolBar를 찾을 수 없습니다.");
+        return FReply::Unhandled();
+    }
+
+    //@Enter 키: ItemSlots로 포커스 이동 (Navigation Down과 동일한 동작)
+    if (Key == EKeys::Enter)
+    {
+        //@Current Item Slots
+        UItemSlots* CurrentItemSlots = Cast<UItemSlots>(GetItemSlotsUI(CurrentItemType));
+        if (!CurrentItemSlots)
+        {
+            UE_LOGFMT(LogInventoryUI, Error, "{0} 타입의 ItemSlots를 찾을 수 없습니다.", *UEnum::GetValueAsString(CurrentItemType));
+            return FReply::Handled();
+        }
+
+        //@첫 번째 아이템 슬롯 유효한가?
+        UInteractableItemSlot* FirstItemSlot = CurrentItemSlots->FindFirstItemSlot();
+        if (!FirstItemSlot || !FirstItemSlot->GetUniqueItemID().IsValid())
+        {
+            UE_LOGFMT(LogInventoryUI, Warning, "{0} 타입의 첫 번째 Item Slot이 유효하지 않습니다.", *UEnum::GetValueAsString(CurrentItemType));
+            return FReply::Handled();
+        }
+
+        //@첫 번째 아이템 슬롯의 강제 호버 상태 전환 요청 이벤트
+        RequestFirstItemSlotHover.Broadcast(CurrentItemType);
+
+        //@SetFocus
+        CurrentItemSlots->SetFocus();
+
+        UE_LOGFMT(LogInventoryUI, Log, "Enter 키: 포커스가 {0} 타입의 ItemSlots로 이동했습니다.", *UEnum::GetValueAsString(CurrentItemType));
+
+        return FReply::Handled();
+    }
+    
+    //@Escape 키: ItemSlots에서 InventoryToolBar로 포커스 복귀
+    if (Key == EKeys::Escape)
+    {
+        //@Current Item Slots
+        UItemSlots* CurrentItemSlots = Cast<UItemSlots>(GetItemSlotsUI(CurrentItemType));
+        if (CurrentItemSlots && CurrentItemSlots->HasKeyboardFocus())
+        {
+            //@Request Cancel Current Hovered Item Slot
+            RequestCancelCurrentHoveredItemSlot.Broadcast(CurrentItemType);
+
+            //@ToolBar에게 포커스 전달
+            if (InventoryToolBar)
+            {
+                InventoryToolBar->SetFocus();
+                UE_LOGFMT(LogInventoryUI, Log, "Escape 키: InventoryToolBar에게 포커스를 전달합니다.");
+                return FReply::Handled();
+            }
+        }
+        else
+        {
+            //@ItemSlots에 포커스가 없으면 상위 위젯에서 처리하도록 Unhandled 반환 및 Focus 해제
+            return FReply::Unhandled().ClearUserFocus();
+        }
+    }
+
+    UE_LOGFMT(LogInventoryUI, Log, "Inventory UI에서 처리하지 않는 키 입력: {0}", *Key.ToString());
+    return FReply::Unhandled();
+}
+
+void UInventoryUI::InternalBindingToInventoryToolBar(UInventoryToolBar* ToolBar)
+{
+    //@Tool Bar
+    if (!ToolBar)
+    {
+        UE_LOGFMT(LogInventoryUI, Error, "ToolBar UI가 유효하지 않습니다.");
+        return;
+    }
+
+    //@내부 바인딩
+    ToolBar->ToolBarInitFinished.BindUFunction(this, "OnInventoryToolBarInitFinished");
+    ToolBar->InventoryToolBarButtonClicked.BindUFunction(this, "OnInventoryToolBarButtonClicked");
+}
+
+void UInventoryUI::InternalBindingToItemSlots(UItemSlots* ItemSlotsWidget)
+{
+    //@Item Slots
+    if (!ItemSlotsWidget)
+    {
+        UE_LOGFMT(LogInventoryUI, Error, "ItemSlots UI가 유효하지 않습니다.");
+        return;
+    }
+    //@초기화 완료 이벤트
+    ItemSlotsWidget->ItemSlotsInitFinished.BindUFunction(this, "OnInventoryItemSlotsInitFinished");
+    ItemSlotsWidget->RequestCancelItemSlotsFocus.BindUFunction(this, "OnRequestCancelItemSlotsFocus");
+}
+
+void UInventoryUI::InternalBindingToItemDescription(UItemDescriptionSlot* ItemDescription)
+{
+    //@Item Description
+    if (!ItemDescription)
+    {
+        UE_LOGFMT(LogInventoryUI, Error, "Item Description UI가 유효하지 않습니다.");
+        return;
+    }
+    //@초기화 완료 이벤트
+    ItemDescription->ItemDescriptionSlotInitFinished.BindUFunction(this, "OnInventoryItemDescriptionInitFinished");
 }
 
 void UInventoryUI::InitializeMenuUIContent()
 {
-
-    //@Create Inventory Content UI
-    CreateInventoryContent();
+    //@Inventory Tool Bar
+    CreateToolBar();
+    //@Item Slots 
+    CreateAllItemSlots();
+    //@Item Description
+    CreateItemDescription();
+    //@초기화 요청 이벤트 호출
+    RequestStartInitByInventoryUI.Broadcast();
 
     //@Super
     Super::InitializeMenuUIContent();
+}
 
+void UInventoryUI::CheckInventoryUIInitialization()
+{
     //@초기화 완료 이벤트 호출
-    RequestStartInitByInventoryUI.Broadcast();
+    if (bInventoryItemSlotsReady && bInventoryToolBarReady && bInventoryItemDescriptionReady)
+    {
+        bInventoryItemSlotsReady = false;
+        bInventoryToolBarReady = false;
+        bInventoryItemDescriptionReady = false;
+
+        for (auto& Pair : MItemSlots)
+        {
+            if (UItemSlots* ItemSlotsWidget = Cast<UItemSlots>(Pair.Value))
+            {
+                RequestFirstItemSlotHover.AddUObject(ItemSlotsWidget, &UItemSlots::OnRequestFirstItemSlotHover);
+                RequestCancelCurrentHoveredItemSlot.AddUObject(ItemSlotsWidget, &UItemSlots::OnRequestCancelCurrentHoveredItemSlot);
+            }
+        }
+
+        //@Reset
+        ResetInventoryUI();
+
+        //@Inventory UI 초기화 완료 체크
+        bInventoryUIContentReady = true;
+        CheckMenuUIContentInitFinished();
+    }
 }
 
 void UInventoryUI::CheckMenuUIContentInitFinished()
@@ -112,71 +329,286 @@ void UInventoryUI::CheckMenuUIContentInitFinished()
 #pragma region SubWidgets
 void UInventoryUI::ResetMenuUIContent()
 {
-    //@Inventory UI Content
-    for (auto& OverlaySlot : InventoryUIContentOverlay->GetAllChildren())
-    {
-        if (auto Widget = Cast<UInventoryUIContent>(OverlaySlot))
-        {
-            Widget->ResetInventoryUIContent();
-            break;
-        }
-    }
-
+    ResetInventoryUI();
     Super::ResetMenuUIContent();
 }
 
-void UInventoryUI::CreateInventoryContent()
+void UInventoryUI::ResetInventoryUI()
 {
-    //@Inventory UI Content 오버레이
-    if (!InventoryUIContentOverlay)
+    UE_LOGFMT(LogInventoryUI, Log, "Inventory UI 구성 위젯들을 초기 상태로 리셋합니다.");
+
+    //@Current Item Type
+    if (CurrentItemType != EItemType::MAX)
     {
-        UE_LOGFMT(LogInventoryUI, Error, "InventoryUIContentOverlay가 유효하지 않습니다.");
-        return;
+        //@Item Slots
+        UUserWidget* ItemSlotsUI = GetItemSlotsUI(CurrentItemType);
+        if (UItemSlots* ItemSlotsWidget = Cast<UItemSlots>(ItemSlotsUI))
+        {
+            ItemSlotsWidget->ResetItemSlots();
+            UE_LOGFMT(LogInventoryUI, Log, "현재 타입({0})의 Item Slots가 초기화되었습니다.", *UEnum::GetValueAsString(CurrentItemType));
+        }
+        else
+        {
+            UE_LOGFMT(LogInventoryUI, Warning, "현재 타입({0})의 Item Slots를 찾을 수 없거나 초기화하지 못했습니다.", *UEnum::GetValueAsString(CurrentItemType));
+        }
+
+        //@Item Description
+        bool bItemDescriptionReset = false;
+        for (auto OverlaySlot : ItemDescriptionOverlay->GetAllChildren())
+        {
+            if (UItemDescriptionSlot* ItemSlotWidget = Cast<UItemDescriptionSlot>(OverlaySlot))
+            {
+                ItemSlotWidget->ResetItemDescriptionSlot();
+                bItemDescriptionReset = true;
+                UE_LOGFMT(LogInventoryUI, Log, "Item Description Slot이 초기화되었습니다.");
+                break;
+            }
+        }
+        if (!bItemDescriptionReset)
+        {
+            UE_LOGFMT(LogInventoryUI, Warning, "Item Description Slot을 찾을 수 없거나 초기화하지 못했습니다.");
+        }
     }
 
-    if (!ensureMsgf(InventoryUIContentClass, TEXT("InventoryUIContentClass가 설정되지 않았습니다.")))
+    //@Inventory Toolbar
+    bool bToolbarReset = false;
+    for (auto OverlaySlot : ToolBarOverlay->GetAllChildren())
+    {
+        UE_LOGFMT(LogInventoryUI, Warning, "{0}", *OverlaySlot->GetName());
+
+        if (UInventoryToolBar* ToolBar = Cast<UInventoryToolBar>(OverlaySlot))
+        {
+            ToolBar->ResetToolBar();
+
+            bToolbarReset = true;
+
+            UE_LOGFMT(LogInventoryUI, Log, "Inventory Toolbar가 초기화되었습니다.");
+            break;
+        }
+    }
+    if (!bToolbarReset)
+    {
+        UE_LOGFMT(LogInventoryUI, Warning, "Inventory Toolbar를 찾을 수 없거나 초기화하지 못했습니다.");
+    }
+
+    UE_LOGFMT(LogInventoryUI, Log, "Inventory UI 구성 위젯들의 초기화가 완료되었습니다.");
+}
+
+void UInventoryUI::CreateToolBar()
+{
+    if (!ensureMsgf(InventoryToolBarClass && ToolBarOverlay, TEXT("InventoryToolBarClass 또는 ToolBarOverlay가 유효하지 않습니다.")))
     {
         return;
     }
-
-    //@Inventory UI Content
-    InventoryUIContent = CreateWidget<UInventoryUIContent>(this, InventoryUIContentClass);
-
-    if (!InventoryUIContent)
+    //@Tool Bar
+    InventoryToolBar = CreateWidget<UInventoryToolBar>(this, InventoryToolBarClass);
+    if (!IsValid(InventoryToolBar))
     {
-        UE_LOGFMT(LogInventoryUI, Error, "Inventory Content UI 위젯 생성에 실패했습니다.");
+        UE_LOGFMT(LogInventoryUI, Error, "InventoryToolBar 위젯 생성에 실패했습니다.");
         return;
     }
-
     //@비동기 초기화 이벤트
-    RequestStartInitByInventoryUI.AddUFunction(InventoryUIContent, "InitializeInventoryUIContent");
-
+    RequestStartInitByInventoryUI.AddUFunction(InventoryToolBar, "InitializeToolBar");
     //@내부 바인딩
-    InternalBindingToInventoryUIContent(InventoryUIContent);
-
-    //@Alignment
-    UOverlaySlot* OverlaySlot = InventoryUIContentOverlay->AddChildToOverlay(InventoryUIContent);
-    if (OverlaySlot)
+    InternalBindingToInventoryToolBar(InventoryToolBar);
+    //@Tool Bar Overlay
+    if (UOverlaySlot* OverlaySlot = ToolBarOverlay->AddChildToOverlay(InventoryToolBar))
     {
         OverlaySlot->SetHorizontalAlignment(HAlign_Fill);
         OverlaySlot->SetVerticalAlignment(VAlign_Fill);
     }
 
-    UE_LOGFMT(LogInventoryUI, Log, "Inventory Content UI 위젯이 성공적으로 추가되었습니다.");
+    UE_LOGFMT(LogInventoryUI, Log, "InventoryToolBar가 성공적으로 생성되고 ToolBarOverlay에 추가되었습니다.");
+}
 
+void UInventoryUI::CreateAllItemSlots()
+{
+    //@Item Slots Overlay
+    if (!IsValid(ItemSlotsOverlay))
+    {
+        UE_LOGFMT(LogInventoryUI, Error, "ItemSlotsOverlay가 유효하지 않습니다.");
+        return;
+    }
+    //@TSet, Item Type 별 하나의 Item Slots만 생성되도록 제한
+    TSet<EItemType> CreatedItemTypes;
+    //@FItemSlotsInfo
+    for (const FItemSlotsInfo& SlotInfo : ItemSlotInformations)
+    {
+        //@Contains
+        if (CreatedItemTypes.Contains(SlotInfo.ItemType))
+        {
+            UE_LOGFMT(LogInventoryUI, Warning, "{0} 타입의 ItemSlots가 이미 생성되었습니다. 중복 생성을 건너뜁니다.", *UEnum::GetValueAsString(SlotInfo.ItemType));
+            continue;
+        }
+        //@Item Slots 블루프린트 클래스
+        if (!SlotInfo.ItemSlotsClass)
+        {
+            UE_LOGFMT(LogInventoryUI, Warning, "{0} 타입의 ItemSlotsClass가 설정되지 않았습니다.", *UEnum::GetValueAsString(SlotInfo.ItemType));
+            continue;
+        }
+        //@Create Widget
+        UItemSlots* NewItemSlots = CreateWidget<UItemSlots>(this, SlotInfo.ItemSlotsClass);
+        if (!NewItemSlots)
+        {
+            UE_LOGFMT(LogInventoryUI, Error, "{0} 타입의 ItemSlots 생성에 실패했습니다.", *UEnum::GetValueAsString(SlotInfo.ItemType));
+            continue;
+        }
+        //@Set Item Type
+        NewItemSlots->SetItemType(SlotInfo.ItemType);
+        //@TMap
+        MItemSlots.Add(SlotInfo.ItemType, NewItemSlots);
+        //@비동기 초기화 이벤트
+        RequestStartInitByInventoryUI.AddUFunction(NewItemSlots, "InitializeItemSlots");
+        //@내부 바인딩
+        InternalBindingToItemSlots(NewItemSlots);
+        //@Item Slots Overlay
+        if (UOverlaySlot* OverlaySlot = ItemSlotsOverlay->AddChildToOverlay(NewItemSlots))
+        {
+            OverlaySlot->SetHorizontalAlignment(HAlign_Fill);
+            OverlaySlot->SetVerticalAlignment(VAlign_Fill);
+        }
+
+        CreatedItemTypes.Add(SlotInfo.ItemType);
+
+        UE_LOGFMT(LogInventoryUI, Log, "{0} 타입의 ItemSlots가 성공적으로 생성되고 ItemSlotsOverlay에 추가되었습니다.",
+            *UEnum::GetValueAsString(SlotInfo.ItemType));
+    }
+
+    //@Debugging
+    for (uint8 i = 0; i < static_cast<uint8>(EItemType::MAX); ++i)
+    {
+        EItemType ItemType = static_cast<EItemType>(i);
+        if (!CreatedItemTypes.Contains(ItemType))
+        {
+            UE_LOGFMT(LogInventoryUI, Warning, "{0} 타입의 ItemSlots가 생성되지 않았습니다.", *UEnum::GetValueAsString(ItemType));
+        }
+    }
+}
+
+void UInventoryUI::CreateItemDescription()
+{
+    //@ItemDescriptionOverlay, ItemDescriptionSlot 블루프린트 클래스
+    if (!ensureMsgf(ItemDescriptionSlotClass && ItemDescriptionOverlay,
+        TEXT("ItemDescriptionSlotClass 또는 ItemDescriptionOverlay가 유효하지 않습니다.")))
+    {
+        UE_LOGFMT(LogInventoryUI, Error, "아이템 설명 UI 생성에 필요한 클래스 또는 오버레이가 설정되지 않았습니다.");
+        return;
+    }
+
+    //@Create Widget
+    UItemDescriptionSlot* ItemDescription = CreateWidget<UItemDescriptionSlot>(this, ItemDescriptionSlotClass);
+    if (!IsValid(ItemDescription))
+    {
+        UE_LOGFMT(LogInventoryUI, Error, "ItemDescriptionSlot 위젯 생성에 실패했습니다.");
+        return;
+    }
+
+    //@비동기 초기화 이벤트에 바인딩
+    RequestStartInitByInventoryUI.AddUFunction(ItemDescription, "InitializeItemDescriptionSlot");
+
+    //@Item Slots의 바인딩 준비 완료 이벤트에 바인딩(Item Description에서 직접적으로 가져오기가 빡셈)
+    ItemSlotsReadyForBinding.AddUFunction(ItemDescription, "OnItemSlotsReadyForBinding");
+
+    //@내부 바인딩
+    InternalBindingToItemDescription(ItemDescription);
+
+    //@Add Child To Overlay
+    UOverlaySlot* OverlaySlot = ItemDescriptionOverlay->AddChildToOverlay(ItemDescription);
+    if (!OverlaySlot)
+    {
+        UE_LOGFMT(LogInventoryUI, Error, "ItemDescriptionSlot을 ItemDescriptionOverlay에 추가하는데 실패했습니다.");
+        return;
+    }
+
+    //@Alignment
+    OverlaySlot->SetHorizontalAlignment(HAlign_Fill);
+    OverlaySlot->SetVerticalAlignment(VAlign_Fill);
+
+    UE_LOGFMT(LogInventoryUI, Log, "ItemDescriptionSlot이 성공적으로 생성되고 ItemDescriptionOverlay에 추가되었습니다.");
+}
+
+void UInventoryUI::UpdateAllItemSlotsVisibility()
+{
+    for (const auto& Pair : MItemSlots)
+    {
+        //@Visibility
+        SetItemTypeVisibility(Pair.Key, Pair.Key == CurrentItemType);
+    }
+}
+
+void UInventoryUI::SetItemTypeVisibility(EItemType ItemType, bool bVisible)
+{
+    //@Visibility: SelfHitTest 나타내기, Collapsed로 숨기기
+    if (UUserWidget* Widget = GetItemSlotsUI(ItemType))
+    {
+        Widget->SetVisibility(bVisible ? ESlateVisibility::SelfHitTestInvisible : ESlateVisibility::Collapsed);
+    }
 }
 #pragma endregion
 
 //@Callbacks
 #pragma region Callbacks
-void UInventoryUI::OnInventoryUIContentInitFinished()
+void UInventoryUI::OnInventoryToolBarInitFinished()
 {
-    UE_LOGFMT(LogInventoryUI, Log, "InventoryUIContent 초기화가 완료되었습니다.");
+    bInventoryToolBarReady = true;
 
-    //@Init Finished
-    bInventoryUIContentReady = true;
-    //@Inventory UI의 초기화 완료 여부 체크
-    CheckMenuUIContentInitFinished();
+    CheckInventoryUIInitialization();
+}
+
+void UInventoryUI::OnInventoryItemSlotsInitFinished()
+{
+    // 모든 ItemSlots가 초기화되었는지 확인
+    bInventoryItemSlotsReady = true;
+
+    CheckInventoryUIInitialization();
+
+    // ItemSlotsOverlay에서 모든 ItemSlots 위젯 찾기
+    for (UWidget* Child : ItemSlotsOverlay->GetAllChildren())
+    {
+        if (UItemSlots* ItemSlotsWidget = Cast<UItemSlots>(Child))
+        {
+            ItemSlotsReadyForBinding.Broadcast(this);
+            break;
+        }
+    }
+}
+
+void UInventoryUI::OnInventoryItemDescriptionInitFinished()
+{
+    //@bInventoryItemDescriptionReady
+    bInventoryItemDescriptionReady = true;
+
+    //@Check 함수
+    CheckInventoryUIInitialization();
+}
+
+void UInventoryUI::OnRequestCancelItemSlotsFocus()
+{
+    //@Set Focus
+    SetFocus();
+}
+
+void UInventoryUI::OnInventoryToolBarButtonClicked(EItemType ItemType)
+{
+    //@EItemType
+    if (CurrentItemType == ItemType)
+    {
+        return;
+    }
+
+    //@Current Item Type
+    CurrentItemType = ItemType;
+
+    //@Update Item Slots Visibility
+    UpdateAllItemSlotsVisibility();
+
+    UE_LOGFMT(LogInventoryUI, Log, "아이템 타입이 {0}(으)로 변경되었습니다.",
+        *UEnum::GetValueAsString(ItemType));
+
+    //@Set Focus
+    SetFocus();
+
+    // TODO: 필요한 경우 추가 로직 구현
 }
 
 void UInventoryUI::OnUIVisibilityChanged_Implementation(ESlateVisibility VisibilityType)
@@ -190,26 +622,57 @@ void UInventoryUI::OnUIVisibilityChanged_Implementation(ESlateVisibility Visibil
     }
 
     //@TODO: Animation 관련 작업 시 해당 함수 오버라이딩...
-
 }
 #pragma endregion
 
 //@Utility(Setter, Getter,...etc)
 #pragma region Utility
-UItemSlots* UInventoryUI::GetItemSlotsByType(EItemType ItemType) const
+UUserWidget* UInventoryUI::GetItemSlotsUI(EItemType ItemType) const
 {
-    if (!InventoryUIContent.Get())
+    // 기존 맵에서 찾기 시도
+    if (auto FoundWidget = MItemSlots.Find(ItemType))
     {
-        UE_LOGFMT(LogInventoryUI, Error, "Inventory UI 내부 컨텐츠가 유효하지 않습니다.");
+        return *FoundWidget;
+    }
+
+    // SpecUp 타입에 대한 특별한 처리
+    if (ItemType == EItemType::SpecUp)
+    {
+        UE_LOGFMT(LogInventoryUI, Warning,
+            "SpecUp 타입의 ItemSlots가 설정되지 않았습니다. 이는 ItemSlotInformations 배열에 SpecUp 설정이 누락되었기 때문일 수 있습니다.");
+
+        // TODO: 향후 SpecUp 전용 UI가 필요하다면 여기서 동적 생성 고려
         return nullptr;
     }
 
-    if (InventoryUIContent)
+    // 다른 타입들에 대한 일반적인 경고
+    UE_LOGFMT(LogInventoryUI, Warning,
+        "{0} 타입의 ItemSlots를 찾을 수 없습니다. ItemSlotInformations에서 해당 타입이 설정되었는지 확인하세요.",
+        *UEnum::GetValueAsString(ItemType));
+
+    return nullptr;
+}
+
+TArray<UItemSlots*> UInventoryUI::GetAllItemTypesItemSlots() const
+{
+    TArray<UItemSlots*> AllItemSlots;
+
+    for (const auto& Pair : MItemSlots)
     {
-        if (UItemSlots* ItemSlotsWidget = CastChecked<UItemSlots>(InventoryUIContent->GetItemSlotsUI(ItemType)))
+        if (UItemSlots* ItemSlots = Cast<UItemSlots>(Pair.Value))
         {
-            return ItemSlotsWidget;
+            AllItemSlots.Add(ItemSlots);
         }
+    }
+
+    return AllItemSlots;
+}
+
+UItemSlots* UInventoryUI::GetItemSlotsByType(EItemType ItemType) const
+{
+    if (UItemSlots* ItemSlotsWidget = CastChecked<UItemSlots>(GetItemSlotsUI(ItemType)))
+    {
+        return ItemSlotsWidget;
     }
 
     return nullptr;
