+++ Source/AgeOfWolves/08_UI/HorizontalDotGauge.cpp
@@ -1,6 +1,3 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-
 #include "HorizontalDotGauge.h"
 #include "Logging/StructuredLog.h"
