+++ Source/AgeOfWolves/14_Subsystem/InteractionManagerSubsystem.h
@@ -1,4 +1,3 @@
-// InteractionManagerSubsystem.h
 #pragma once
 
 #include "CoreMinimal.h"
